06_django_quiz/requirements-dev.txt
-r requirements.txt

django-extensions
flake8
flake8-builtins
flake8-import-order
ipython
pep8-naming
